Changed the way the STOP command works.
This is because TWS 944 appears to have changed the menu structure so that although to the user it looks like there is a File > Exit menu item, in fact there isn't. There IS a File > Close menu item, but programmatically clicking that that doesn't actually do anything. Consequently in TWS 944 the STOP command, and the tidy exit configured by the ClosedownAt setting, don't work.

The new approach is to send a WindowEvent.WINDOW_CLOSING message on the AWT system event queue, which is actually a lot simpler.