Allow TWS settings directory to be defaulted.
IbDir was a required setting in IBController.ini, Since it is usually the same as the directory in which TWS was installed, it is unnecessary to make it mandatory. Making it optional means that the sample IBController.ini supplied in the distribution ZIP is platform independent, since it no longer has any Windows-specific path syntax.

It now defaults to the current working directory when IBController is started.