Modify build to get IBController.jar into resources folder
Currently the IBController.jar file is built into an ignored folder,
which means the copy of IBController.jar in resources needs to be moved
there manually. This hasn't been happening, so anyone who clones the
IBController repository has been getting an out-of-date jar file. Now
the build is changed to create IBController.jar in resources, and copy
it from there to create the .zip file.