Add /INLINE argument for use with Task Scheduler
This ensures that the window that contains the information messages and
that acts as the Java console is what is directly started by Task
Scheduler. It's only this window whose termination terminates the task:
since closing TWS/Gateway closes this window, it thus also terminates
the task.

See #130 for further background.

Note that this does nothing to address the problem on Windows 10
detailed in #138.