Remove unnecessary statement
Fixes issue #154.
The previous line (line 198) clicks the OK button which closes and
destroys the dialog object. The close event causes the configDialog
variable to be set to null when the GlobalConfigurationDialogHandler
calls clearConfigDialog(). Since this happens on a different thread, a
race error woud sometimes occur, with the configDialog variable being
set to null before line 199 was executed: thus NullPointerException.